// client/src/types/UserBodyMetricsPayload.ts
export type UserBodyMetricsPayload = {
    age: number;
    gender: string;
    height: number;
    weight: number;
    fitnessGoal: string;
    targetWeight: number;
    dietPreference: string;
    experienceLevel: string;
    workoutLocation: string;
    preferredWorkoutTypes: string[];
    medicalConditions: string[];
    profileCompleted: boolean;
};